Migrate to command pattern (#55637)
