+++ internal/database/database.go
@@ -3,7 +3,6 @@ package database
 import (
 	"context"
 	"database/sql"
-	"fmt"
 	"time"
 
 	"github.com/sourcegraph/log"
@@ -15,67 +14,28 @@ import (
 	gha "github.com/sourcegraph/sourcegraph/internal/github_apps/store"
 )
 
-// ExecuteWithClient executes a DBExecutable with the provided DBClient.
-// The client executes the Executable using the ExecuteRaw method, which
-// is then immediately converted by the readResponse function.
-func ExecuteWithClient[T any](ctx context.Context, dbe DBExecutable[T], client DBClient) (T, error) {
-	resp, err := client.Execute(ctx, dbe)
-
-	respT, ok := resp.(T)
-	if !ok {
-		var t T
-		panic(fmt.Sprintf(`Return type of %[1]T.Execute does not match return type of %[1]T.ExecuteRaw.
-
-The implementation for DBExecutable[%[2]T] should match the following:
-
-func (q %[1]T) Execute(ctx context.Context, store *basestore.Store) (%[2]T, error) {...}
-
-func (q %[1]T) ExecuteRaw(ctx context.Context, store *basestore.Store) (any, error) {
-	return q.Execute(ctx, store)
-}`, dbe, t))
-	}
-
-	return respT, err
+// A DBCommand represents a contained series of database interactions that
+// serves a singular purpose (i.e. fetch all users from the database).
+//
+// A DBCommand requires a context and a store to execute, and any results
+// are contained within the command itself. Once a command has been executed,
+// the results can be read.
+type DBCommand interface {
+	Execute(context.Context, *basestore.Store) error
 }
 
-// A DBExecutableRaw uses a store to perform various operations on a database.
-//
-// Execute returns an `any` type, and it is the responsibility of the caller
-// to ensure that the response is casted to the correct response type.
-//
-// This is not meant to be used outside of a dedicated Client implementation.
-//
-// An Execute call can require multiple database calls to produce its result.
-//
-// See ReadResponse for a method to conveniently process an Execute response.
-type DBExecutableRaw interface {
-	ExecuteRaw(context.Context, *basestore.Store) (any, error)
+// A DBStore is capable of executing DBCommands. It provides the DBCommand
+// with the context and store it requires to execute.
+type DBStore interface {
+	ExecuteCommand(context.Context, DBCommand) error
 }
 
-// A DBExecutable uses a store to perform various operations on a database.
-//
-// An Execute call can require multiple database calls to produce its result.
-type DBExecutable[T any] interface {
-	DBExecutableRaw
-	Execute(context.Context, *basestore.Store) (T, error)
-}
-
-// A DBClient is a database client capable of executing DBQueries.
-//
-// It is higher-level than a basestore.Store, and will most likely be passing
-// a basestore.Store to the DBQuery in order to execute it.
-//
-// An Execute call can lead to multiple database queries.
-type DBClient interface {
-	Execute(context.Context, DBExecutableRaw) (any, error)
-}
-
-type dbClient struct {
+type dbStore struct {
 	store *basestore.Store
 }
 
-func (c *dbClient) Execute(ctx context.Context, query DBExecutableRaw) (any, error) {
-	return query.ExecuteRaw(ctx, c.store)
+func (c *dbStore) ExecuteCommand(ctx context.Context, command DBCommand) error {
+	return command.Execute(ctx, c.store)
 }
 
 // DB is an interface that embeds dbutil.DB, adding methods to
@@ -86,7 +46,7 @@ type DB interface {
 	dbutil.DB
 	basestore.ShareableStore
 
-	Client() DBClient
+	DBStore() DBStore
 	AccessTokens() AccessTokenStore
 	Authz() AuthzStore
 	BitbucketProjectPermissions() BitbucketProjectPermissionsStore
@@ -197,8 +157,8 @@ func (d *db) Done(err error) error {
 	return d.Store.Done(err)
 }
 
-func (d *db) Client() DBClient {
-	return &dbClient{
+func (d *db) DBStore() DBStore {
+	return &dbStore{
 		store: d.Store,
 	}
 }
