+++ internal/database/mocks_temp.go
@@ -12750,7 +12750,7 @@ func NewMockDB() *MockDB {
 			},
 		},
 		ClientFunc: &DBClientFunc{
-			defaultHook: func() (r0 DBClient) {
+			defaultHook: func() (r0 DBStore) {
 				return
 			},
 		},
@@ -13082,7 +13082,7 @@ func NewStrictMockDB() *MockDB {
 			},
 		},
 		ClientFunc: &DBClientFunc{
-			defaultHook: func() DBClient {
+			defaultHook: func() DBStore {
 				panic("unexpected invocation of MockDB.Client")
 			},
 		},
@@ -13404,7 +13404,7 @@ func NewMockDBFrom(i DB) *MockDB {
 			defaultHook: i.BitbucketProjectPermissions,
 		},
 		ClientFunc: &DBClientFunc{
-			defaultHook: i.Client,
+			defaultHook: i.DBStore,
 		},
 		CodeHostsFunc: &DBCodeHostsFunc{
 			defaultHook: i.CodeHosts,
@@ -14083,15 +14083,15 @@ func (c DBBitbucketProjectPermissionsFuncCall) Results() []interface{} {
 // DBClientFunc describes the behavior when the Client method of the parent
 // MockDB instance is invoked.
 type DBClientFunc struct {
-	defaultHook func() DBClient
-	hooks       []func() DBClient
+	defaultHook func() DBStore
+	hooks       []func() DBStore
 	history     []DBClientFuncCall
 	mutex       sync.Mutex
 }
 
-// Client delegates to the next hook function in the queue and stores the
+// DBStore delegates to the next hook function in the queue and stores the
 // parameter and result values of this invocation.
-func (m *MockDB) Client() DBClient {
+func (m *MockDB) DBStore() DBStore {
 	r0 := m.ClientFunc.nextHook()()
 	m.ClientFunc.appendCall(DBClientFuncCall{r0})
 	return r0
@@ -14099,7 +14099,7 @@ func (m *MockDB) Client() DBClient {
 
 // SetDefaultHook sets function that is called when the Client method of the
 // parent MockDB instance is invoked and the hook queue is empty.
-func (f *DBClientFunc) SetDefaultHook(hook func() DBClient) {
+func (f *DBClientFunc) SetDefaultHook(hook func() DBStore) {
 	f.defaultHook = hook
 }
 
@@ -14107,7 +14107,7 @@ func (f *DBClientFunc) SetDefaultHook(hook func() DBClient) {
 // Client method of the parent MockDB instance invokes the hook at the front
 // of the queue and discards it. After the queue is empty, the default hook
 // function is invoked for any future action.
-func (f *DBClientFunc) PushHook(hook func() DBClient) {
+func (f *DBClientFunc) PushHook(hook func() DBStore) {
 	f.mutex.Lock()
 	f.hooks = append(f.hooks, hook)
 	f.mutex.Unlock()
@@ -14115,20 +14115,20 @@ func (f *DBClientFunc) PushHook(hook func() DBClient) {
 
 // SetDefaultReturn calls SetDefaultHook with a function that returns the
 // given values.
-func (f *DBClientFunc) SetDefaultReturn(r0 DBClient) {
-	f.SetDefaultHook(func() DBClient {
+func (f *DBClientFunc) SetDefaultReturn(r0 DBStore) {
+	f.SetDefaultHook(func() DBStore {
 		return r0
 	})
 }
 
 // PushReturn calls PushHook with a function that returns the given values.
-func (f *DBClientFunc) PushReturn(r0 DBClient) {
-	f.PushHook(func() DBClient {
+func (f *DBClientFunc) PushReturn(r0 DBStore) {
+	f.PushHook(func() DBStore {
 		return r0
 	})
 }
 
-func (f *DBClientFunc) nextHook() func() DBClient {
+func (f *DBClientFunc) nextHook() func() DBStore {
 	f.mutex.Lock()
 	defer f.mutex.Unlock()
 
@@ -14163,7 +14163,7 @@ func (f *DBClientFunc) History() []DBClientFuncCall {
 type DBClientFuncCall struct {
 	// Result0 is the value of the 1st result returned from this method
 	// invocation.
-	Result0 DBClient
+	Result0 DBStore
 }
 
 // Args returns an interface slice containing the arguments of this
